// src/contextualHelp/ui/common/redux/reducers/types.ts
export interface ICellAction {
    cellId: string | undefined;
}
// ...
export interface ICellAndCursorAction extends ICellAction {
    cursorPos: CursorPos;
}